Added, Moved and renamed files
added few files to the root, movded ewr_departures_UA to root,
moved pkl files to root, fixed error prone to scheduld and actual data
format. Reason for error date and time format had to deal with 'Not
available' and so added that to the root of the scheduled_flights()
method when filtering out reliable flights. Initiated python file for
extracting departure and destination of a flight number. need more work.